reduced casting, no more conversion warnings

+++ SpaceInvaders_2023/Source/Helper.cpp
@@ -4,7 +4,7 @@
 
 float lineLength(Vector2 A, Vector2 B) noexcept
 {
-	return sqrtf(pow(B.x - A.x, 2) + pow(B.y - A.y, 2));
+	return sqrtf(static_cast<float>(pow(B.x - A.x, 2) + pow(B.y - A.y, 2)));
 }
 
 static bool pointInCircle(Vector2 circlePos, float radius, Vector2 point) noexcept
@@ -19,9 +19,9 @@ bool CheckCollision(Vector2 circleOrigin, float circleRadius, Vector2 A, Vector2
 		return true;
 	}
 
-	constexpr float buffer = 0.1;
+	constexpr float buffer = 0.1f;
 	const float length = lineLength(A, B);
-	const float dotP = (((circleOrigin.x - A.x) * (B.x - A.x)) + ((circleOrigin.y - A.y) * (B.y - A.y))) / pow(length, 2);
+	const float dotP = static_cast<float>((((circleOrigin.x - A.x) * (B.x - A.x)) + ((circleOrigin.y - A.y) * (B.y - A.y))) / pow(length, 2));
 
 	const float closestX = A.x + (dotP * (B.x - A.x));
 	const float closestY = A.y + (dotP * (B.y - A.y));
